+++ src/app/explore/explore.component.ts
@@ -3,6 +3,7 @@ import { Validators, FormGroup, FormBuilder } from '@angular/forms';
 import { EtherscanService } from '../services/etherscan.service';
 import { Event } from '../models/event';
 import { EventResponse } from '../models/event-response';
+import { LeaderBoard } from '../models/leader-board';
 import { Colors } from '../models/colors';
 
 @Component({
@@ -14,6 +15,7 @@ import { Colors } from '../models/colors';
 export class ExploreComponent implements OnInit {
   events: Event[];
   eventData: EventResponse[];
+  lb: LeaderBoard = new LeaderBoard();
   colors: Colors = new Colors();
   txRegex: string = "0x[a-z]{64}";
   eventForm: FormGroup;
@@ -27,6 +29,7 @@ export class ExploreComponent implements OnInit {
       this.events = events;
     });
     this.createFormGroups();
+    this.initLeaderBoard();
   }
 
   createFormGroups() {
@@ -45,6 +48,58 @@ export class ExploreComponent implements OnInit {
     return this.nftForm.hasError('required', [value]) ? 'Required' : '';
   }
 
+  initLeaderBoard() {
+    let _data: Array < any > = [this.events[0].fromBlock, this.events[0].address, this.events[0].topic];
+    let data: Array < any > = [this.events[1].fromBlock, this.events[1].address, this.events[1].topic];
+    let donorObj = {};
+    let nonprofitObj = {};
+
+    this.ess.getEventData(data).subscribe(events => {
+        let result = this.ess.processDonationReceived(events);
+        for (var i = 0; i < result.length; i++) {
+            let valueForKey = result[i].donor in donorObj;
+            if(!valueForKey){
+              donorObj[result[i].donor] = result[i].amount;
+            } else {
+              donorObj[result[i].donor] += result[i].amount;
+            }
+        }
+        console.log(donorObj);
+        this.setLeaderBoard(donorObj, true);
+    });
+    this.ess.getEventData(_data).subscribe(events => {
+        let result = this.ess.processPaymentDistributed(events);
+        for (var i = 0; i < result.length; i++) {
+            let valueForKey = result[i].to in nonprofitObj;
+            if(!valueForKey){
+              nonprofitObj[result[i].to] = result[i].amount;
+            } else {
+              nonprofitObj[result[i].to] += result[i].amount;
+            }
+        }
+        console.log(nonprofitObj);
+        this.setLeaderBoard(nonprofitObj, false);
+    });
+  }
+
+  setLeaderBoard(obj: Object, state: boolean){
+    if(state){
+        this.lb.donor = [];
+        this.lb.donorAmount = [];
+        for (let [key, value] of Object.entries(obj)) {
+            this.lb.donor.push(key);
+            this.lb.donorAmount.push(Number(value));
+        }
+    } else {
+        this.lb.nonprofit = [];
+        this.lb.nonprofitAmount = [];
+        for (let [key, value] of Object.entries(obj)) {
+            this.lb.nonprofit.push(key);
+            this.lb.nonprofitAmount.push(Number(value));
+        }
+    }
+  }
+
   onEventSelected(index): void {
     let data: Array < any > = [this.events[index].fromBlock, this.events[index].address, this.events[index].topic];
     this.ess.getEventData(data).subscribe(events => {
@@ -59,7 +114,6 @@ export class ExploreComponent implements OnInit {
           console.log('error processing event data');
           break;
       }
-      console.log(this.eventData);
     });
   }