+++ src/app/models/leader-board.ts
@@ -1,0 +1,6 @@
+export class LeaderBoard {
+	donor: Array<string>;
+	donorAmount: Array<Number>;
+    nonprofit: Array<string>;
+    nonprofitAmount: Array<Number>;
+}