+++ src/app/services/caching-interceptor.service.ts
@@ -1,0 +1,31 @@
+import { Injectable } from '@angular/core';
+import { HttpEvent, HttpRequest, HttpResponse, HttpInterceptor, HttpHandler } from '@angular/common/http';
+
+import { Observable, of } from 'rxjs';
+import { startWith, tap } from 'rxjs/operators';
+//import 'rxjs/add/observable/of';
+
+import { RequestCache } from './caching-request.service';
+
+@Injectable()
+export class CachingInterceptor implements HttpInterceptor {
+  constructor(private cache: RequestCache) {}
+
+  intercept(req: HttpRequest<any>, next: HttpHandler) {
+    const cachedResponse = this.cache.get(req);
+    return cachedResponse ? of(cachedResponse) : this.sendRequest(req, next, this.cache);
+  }
+
+  sendRequest(
+    req: HttpRequest<any>,
+    next: HttpHandler,
+    cache: RequestCache): Observable<HttpEvent<any>> {
+    return next.handle(req).pipe(
+      tap(event => {
+        if (event instanceof HttpResponse) {
+          cache.put(req, event);
+        }
+      })
+    );
+  }
+}